Fix Vulkan renderer warnings and add r_swapInterval support
Fix all build warnings in the Vulkan renderer (shadow variables, format
specifiers, strict prototypes, array size mismatches, header guard typo,
unhandled switch cases, and a real memcpy bug in R_MovePatchSurfacesToHunk
where heightLodError was copied to itself instead of the hunk allocation).

Add r_swapInterval cvar support so the Vulkan renderer respects vsync
settings. Previously the present mode selection always preferred MAILBOX
over FIFO regardless of r_swapInterval, causing vsync to have no effect
and frame rates to run uncapped at 1000+ fps.

+++ code/renderer_vulkan/vk_image.c
@@ -123,8 +123,8 @@ static void vk_createStagingBuffer(uint32_t size)
 
         VK_CHECK(qvkBindBufferMemory(vk.device, StagBuf.buff, StagBuf.mappableMem, 0));
 
-        ri.Printf(PRINT_ALL, " Stagging buffer alignment: %ld, memoryTypeBits: 0x%x, Type Index: %d. \n",
-            memory_requirements.alignment, memory_requirements.memoryTypeBits, alloc_info.memoryTypeIndex);
+        ri.Printf(PRINT_ALL, " Stagging buffer alignment: %llu, memoryTypeBits: 0x%x, Type Index: %d. \n",
+            (unsigned long long)memory_requirements.alignment, memory_requirements.memoryTypeBits, alloc_info.memoryTypeIndex);
     }
 }
 
@@ -333,8 +333,8 @@ static void vk_createImageAndBindWithMemory(image_t* pImg)
 
     ri.Printf(PRINT_ALL, " --- Device memory allocation --- \n");
 
-    ri.Printf(PRINT_ALL, "alignment: %ld, Type Index: %d. \n",
-            memory_requirements.alignment, alloc_info.memoryTypeIndex);
+    ri.Printf(PRINT_ALL, "alignment: %llu, Type Index: %d. \n",
+            (unsigned long long)memory_requirements.alignment, alloc_info.memoryTypeIndex);
     
     ri.Printf(PRINT_ALL, "Image chuck memory consumed: %d M \n",
             devMemImg.Index * (IMAGE_CHUNK_SIZE >> 20) );
